add mapper

+++ src/main/java/com/tech/app/domain/Category.java
@@ -18,7 +18,7 @@ import javax.persistence.Id;
 public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idCategory;
+    private long idCategory;
 
     private String name;
 }
