+++ src/main/java/com/tech/app/service/dto/CategoryDTO.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class CategoryDTO {
-    private int idCategory;
+    private long idCategory;
     private String name;
 }
